test/fixtures/python-fixture/main.py
import asyncio
from typing import Optional
from ztvs_sdk import Metadata, run, Check, Finding # type: ignore

class PythonCheck(Check):
    def id(self) -> str:
        return "python_test_check"

    def name(self) -> str:
        return "Python Fixture Check"

    async def run(self) -> Optional[Finding]:
        return Finding(
            id="F-PY-001",
            severity="info",
            title="Python Fixture Executed",
            description="This is a successful result from the Python polyglot fixture.",
            evidence={"runtime": "python3"},
            remediation="No action required."
        )

async def main():
    meta = Metadata(
        name="python-fixture",
        version="1.0.0",
        api_version=1
    )
    await run(meta, [PythonCheck()])

if __name__ == "__main__":
    asyncio.run(main())
