test/fixtures/python-fixture/requirements.txt
ztvs-sdk-python @ git+https://github.com/mosesgameli/ztvs-sdk-python.git
pydantic>=2.0.0
